requirements.txt
pymongo
python-dotenv
passlib==1.7.4
bcrypt==4.1.3
fastapi
uvicorn[standard]
pydantic
certifi
